add spinner to TransactionList, refactor heading to separate cmp,  show txs graph while fetching next page
