+++ packages/suite/src/views/wallet/transactions/components/TransactionList/index.tsx
@@ -1,14 +1,13 @@
-/* eslint-disable radix */
 import React, { useMemo } from 'react';
-import { FormattedDate } from 'react-intl';
 import styled from 'styled-components';
-import { colors, variables } from '@trezor/components';
+import { colors, variables, Loader } from '@trezor/components';
 import { groupTransactionsByDate, parseKey, sumTransactions } from '@wallet-utils/transactionUtils';
 import { SETTINGS } from '@suite-config';
 import { Account, WalletAccountTransaction } from '@wallet-types';
 import TransactionItem from './components/TransactionItem/Container';
 import Pagination from './components/Pagination';
-import { Card, FiatValue, HiddenPlaceholder, Translation } from '@suite-components';
+import { Card, Translation } from '@suite-components';
+import TransactionHeaderItem from './components/TransactionHeaderItem';
 
 const Wrapper = styled.div``;
 
@@ -21,54 +20,23 @@ const Transactions = styled.div`
     flex-direction: column;
 `;
 
-const DayHeading = styled.div`
-    position: sticky;
-    top: 0;
-    display: flex;
-    font-size: ${variables.FONT_SIZE.TINY};
-    min-height: 35px; /* same as height of badge with fiat value plus padding */
-    color: ${colors.BLACK50};
-    border-bottom: 2px solid ${colors.BLACK96};
-    font-weight: ${variables.FONT_WEIGHT.DEMI_BOLD};
-    padding: 5px;
-    text-transform: uppercase;
-    background: ${colors.WHITE};
-    justify-content: space-between;
-    align-items: center;
-
-    &:first-child {
-        padding-top: 0px;
-        border-top-left-radius: 6px;
-        border-top-right-radius: 6px;
-    }
-`;
-
 const PaginationWrapper = styled.div`
     margin: 10px 0px;
 `;
 
-const DayAmountWrapper = styled.div`
+const LoaderWrapper = styled.div`
     display: flex;
+    flex-direction: column;
+    justify-items: center;
     align-items: center;
+    margin: 24px 0px;
 `;
 
-const DayAmount = styled.div`
+const LoaderText = styled.div`
     display: flex;
-
-    & + & {
-        margin-left: 14px;
-    }
-`;
-
-const FiatDayAmount = styled(DayAmount)`
-    min-width: 100px;
-    justify-content: flex-end;
-    text-align: right;
-    margin-left: 16px;
-`;
-
-const DateWrapper = styled.div`
-    display: flex;
+    color: ${colors.BLACK25};
+    font-size: ${variables.FONT_SIZE.SMALL};
+    text-align: center;
 `;
 
 interface Props {
@@ -78,6 +46,7 @@ interface Props {
     totalPages?: number;
     perPage: number;
     symbol: Account['symbol'];
+    isLoading?: boolean;
     onPageSelected: (page: number) => void;
 }
 
@@ -88,6 +57,7 @@ const TransactionList = ({
     totalPages,
     onPageSelected,
     perPage,
+    isLoading,
     ...props
 }: Props) => {
     const startIndex = (currentPage - 1) * perPage;
@@ -109,63 +79,39 @@ const TransactionList = ({
     const isOnLastPage = slicedTransactions.length < SETTINGS.TXS_PER_PAGE;
     const shouldShowRipplePagination = !(currentPage === 1 && isOnLastPage);
     const showPagination = totalPages ? totalPages > 1 : shouldShowRipplePagination;
+
     return (
         <Wrapper>
             <StyledCard>
                 <Transactions>
-                    {Object.keys(transactionsByDate).map(dateKey => {
-                        const parsedDate = parseKey(dateKey);
-                        const totalAmountPerDay = sumTransactions(transactionsByDate[dateKey]);
-                        return (
-                            <React.Fragment key={dateKey}>
-                                <DayHeading>
-                                    {dateKey === 'pending' ? (
-                                        <DateWrapper>
-                                            <Translation id="TR_PENDING" />
-                                        </DateWrapper>
-                                    ) : (
-                                        <>
-                                            <DateWrapper>
-                                                <FormattedDate
-                                                    value={parsedDate ?? undefined}
-                                                    day="numeric"
-                                                    month="long"
-                                                    year="numeric"
-                                                />
-                                            </DateWrapper>
-                                            <DayAmountWrapper>
-                                                <HiddenPlaceholder>
-                                                    <DayAmount>
-                                                        {totalAmountPerDay.gte(0) && '+'}
-                                                        {totalAmountPerDay.toFixed()}{' '}
-                                                        {props.symbol.toUpperCase()}
-                                                    </DayAmount>
-                                                </HiddenPlaceholder>
-                                                {/* TODO: daily deltas are multiplied by current rate instead of the rate for given day. if someone notices. calc average rate of all txs in a day and use that? */}
-                                                <HiddenPlaceholder>
-                                                    <FiatValue
-                                                        amount={totalAmountPerDay.toFixed()}
-                                                        symbol={props.symbol}
-                                                    >
-                                                        {({ value }) =>
-                                                            value && (
-                                                                <FiatDayAmount>
-                                                                    {value}
-                                                                </FiatDayAmount>
-                                                            )
-                                                        }
-                                                    </FiatValue>
-                                                </HiddenPlaceholder>
-                                            </DayAmountWrapper>
-                                        </>
+                    {isLoading ? (
+                        <LoaderWrapper>
+                            <Loader size={28} />
+                            <LoaderText>
+                                <Translation id="TR_LOADING_TRANSACTIONS" />
+                            </LoaderText>
+                        </LoaderWrapper>
+                    ) : (
+                        Object.keys(transactionsByDate).map(dateKey => {
+                            const parsedDate = parseKey(dateKey);
+                            const totalAmountPerDay = sumTransactions(transactionsByDate[dateKey]);
+                            return (
+                                <React.Fragment key={dateKey}>
+                                    <TransactionHeaderItem
+                                        dateKey={dateKey}
+                                        parsedDate={parsedDate}
+                                        totalAmountPerDay={totalAmountPerDay}
+                                        symbol={props.symbol}
+                                    />
+                                    {transactionsByDate[dateKey].map(
+                                        (tx: WalletAccountTransaction) => (
+                                            <TransactionItem key={tx.txid} transaction={tx} />
+                                        ),
                                     )}
-                                </DayHeading>
-                                {transactionsByDate[dateKey].map((tx: WalletAccountTransaction) => (
-                                    <TransactionItem key={tx.txid} transaction={tx} />
-                                ))}
-                            </React.Fragment>
-                        );
-                    })}
+                                </React.Fragment>
+                            );
+                        })
+                    )}
                 </Transactions>
                 {showPagination && (
                     <PaginationWrapper>
